# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auth api",
        "description": "包含功能：API：登录，注册，验证码   后台：带权限，客户管理。  header：版本号v,来源origin 认证auth, time ",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1/api"
        }
    ],
    "paths": {
        "/projects": {
            "get": {
                "operationId": "1cd59119d6cd462cc8e2b071a43e2469",
                "responses": {
                    "200": {
                        "description": "Display a listing of projects."
                    }
                }
            }
        },
        "/getCaptcha": {
            "get": {
                "tags": [
                    "api"
                ],
                "summary": "获得验证码",
                "description": "获得验证码",
                "operationId": "getCaptcha",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "code"
                                    ],
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "key": {
                                                    "type": "string"
                                                },
                                                "img": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "default": ""
                                        },
                                        "code": {
                                            "type": "integer",
                                            "default": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "api"
                ],
                "summary": "注册",
                "description": "注册",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "nickname",
                        "in": "query",
                        "description": "昵称",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "邮箱",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "密码",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "description": "二维码",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "messageId",
                        "in": "query",
                        "description": "获取验证吗返回的messageId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "code"
                                    ],
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "access_token": {
                                                    "type": "string"
                                                },
                                                "token_type": {
                                                    "type": "string"
                                                },
                                                "expires_in": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "default": ""
                                        },
                                        "code": {
                                            "type": "integer",
                                            "default": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "api"
                ],
                "summary": "登录",
                "description": "登录",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "邮箱",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "密码",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "description": "二维码",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "messageId",
                        "in": "query",
                        "description": "获取验证吗返回的messageId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "code"
                                    ],
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "access_token": {
                                                    "type": "string"
                                                },
                                                "token_type": {
                                                    "type": "string"
                                                },
                                                "expires_in": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "default": ""
                                        },
                                        "code": {
                                            "type": "integer",
                                            "default": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "api"
                ],
                "summary": "登出",
                "description": "登出",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "default": ""
                                        },
                                        "msg": {
                                            "type": "string",
                                            "default": "xx not correct"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "default": -1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/resetPwd": {
            "post": {
                "tags": [
                    "api"
                ],
                "summary": "修改密码",
                "description": "修改密码",
                "operationId": "resetPwd",
                "parameters": [
                    {
                        "name": "password",
                        "in": "query",
                        "description": "旧密码",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "newpassword",
                        "in": "query",
                        "description": "新密码",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "code"
                                    ],
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "default": ""
                                        },
                                        "code": {
                                            "type": "integer",
                                            "default": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}